requirements.txt
astroid==2.8.3
autopep8==1.5.7
decorator==5.1.0
flake8==4.0.1
greenlet==1.1.2
importlib-metadata==4.2.0
install==1.3.4
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysql-connector-python==8.0.27
numpy==1.19.5
opencv-contrib-python==3.4.1.15
opencv-python==3.4.1.15
pbr==5.6.0
platformdirs==2.4.0
protobuf==3.19.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.11.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.26
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
Tempita==0.5.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
wrapt==1.13.2
zipp==3.6.0
